extracted details Header to own component
